Redo AC calculations using fixedp
This was using various scale parameters under the hood, and this commit
cleans up these calculations using fixedp.

+++ crawl-ref/source/describe.cc
@@ -1556,7 +1556,7 @@ static string _describe_armour(const item_def &item, bool verbose)
             {
                 description += make_stringf("\nWearing mundane armour of this type "
                                             "will give the following: %d AC",
-                                             you.base_ac_from(item));
+                                             (int) you.base_ac_from(item));
             }
         }
     }
